Support image_type x arch specific upload options
Until now, the JSON file with upload options provided on the CLI was
expected to contain a single object representing the upload options to
be used for all ImageRequests created by the builder plugin for
combinations of provided arches and image_types.

This commit add additional option to specify the upload options in the
provided JSON file as an array of image_type and arch -specific upload
options in the form of:

```
[
	{
		"arch": "<arch_name>",
		"image_type": "<image_type_name>",
		"upload_options": {
			...
		}
	},
	...
]
```

If the array is provided, then the specified upload options will be used
only for the specific arch / image_type ImageRequest. This allows one to
specify different upload options per architecture or image_type or both,
or even specify upload options only for a subset of ImageRequests.

Extend unit tests to cover the newly added functionality.